Now passes reqesusts through to of-timezone
Because we need to change the timezone as an unprivileged user it became obvious that time zone control should really be an OS thing, not a SQP thing.  So the OS has been updated to allow time zone alterations without privilege escalation and this just passes requests to the of-timezone script.